requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic==1.10.18
pandas==2.2.3
openpyxl==3.1.5
python-multipart==0.0.9
redis==5.0.8
celery==5.4.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
httpx==0.27.2
python-multipart